add initial Javascript gameplay for humans
bots are currently not supported, and end-game text not showing up.
Active player always shows as Player 1 even when it's Player 2's turn.
But the game can be played as is. Current issue is that if the window is
smaller than the board, player movement may not align to tiles.