add turnType and color to models, add remove_tile()
turntype allows board to keep track of existing moves, and color will
keep the players consistent.
Problem is, after we make a move, we actually save the board as a new board
rather than saving TO the old board. So the first move works because its
from the same "fresh" board that never gets saved. Subsequent moves fail
because it's attempting to remove a tile on the "fresh" board who's current
turn is moving a player