auto-assign players with cookies
new users will have a user_id assigned to them (based on a uuid), and be
immediately tied to an open user of any game, with preference for the
active player. If the user's game-piece is not active, no links will be
set on the board. Same for a user just observing the match.