display game code and allow friends to join
game code shows above main board. localhost:8000 has two buttons to start
or join a game. You can also join a new game with the uuid you type. A 404
is returned if a clashing uuid is generated by a newly-created game. We
clear all tables when that happens. It won't trigger if a user an existing
game, thankfully