restructure Robots assignment and handling
robot players are preferentially players who go last. MoveBot class is
instantiated everytime we make a robot turn, rather than storing in a dict.
This is a django optimization step because the player model object in
django is not hashable until saved; and the assigning of robots to player
relationship is done before saving.