add robots to django game
robots take their turn when a refresh is called and the game state declares
it's the robots turn. Once the turn is complete the refresh head will
naturally be directed to request the updated board. If any player joins she
will automatically take the place of an open robot on the board. In this
way the game naturally moves from being played by users to being played
by humans. However, this setup means that an observer can't join just to
watch a robot-vs-player game