add players relation to tile
player object must be saved first before tile now, thanks to new relation.
This was a fix for a "rare bug" in that the tile-remover did not correctly
identify that a player occupied a tile. So it would attempt to move to it's
own tile, and somehow successfully "move" in that way.
This bug was because the tile.player property was not always filled in by
pre-processing. making .player a property on tile should automatically fix
this.